+++ packages/juice_realtime/lib/src/use_cases/initialize_realtime_use_case.dart
@@ -10,7 +10,7 @@ class InitializeRealtimeUseCase
   Future<void> execute(InitializeRealtimeEvent event) async {
     bloc.configure(event.config);
     if (event.config.autoConnect) {
-      bloc.send(ConnectEvent());
+      await bloc.send(ConnectEvent());
     }
   }
 }